CSL-232. Add notion of books assigned to periods.
If user is in a Period, then only books that have a BookAssignment to that Period
are shown on their library page.
Guest users and others not in a period see all books.
Very simplistic right now, will be expanded later.
Assignments need to be made through the admin site, there is no user-facing UI yet.