Always look up clusive user and make it available to views and templates.
Rather than doing this in every single view, look it up in a middleware function.
Being attached to the request object, it is also available to templates.
This fixes bug that was causing 'change password' to not be available on many pages.
For CSL-339, use this to make upload unavailable to guests.
Add cancel/back options to upload form.