Make management commands refuse to import a public book that already exists.
There is really no way other than checking in the database for a newly-started docker container to know whether the import has already happened.
In the future, we will want to have a process that can update a book with a revised version.
For now, deleting a book will allow the new version to get uploaded.