CSL-1036: WIP Requesting access tokens for additional scopes.
This version creates the request to Google and attempts to handle
the callback by using plain Django idioms and not use the Google
Python client.  The problem with using the client is that the
parameter string in the callback that Google issues begins
begins with a hash ('#'), making it impossible to acquire
the parameters in the <WSGIRequest: GET '/account/add_scope_callback/'>
object.