CSL-1036: WIP Requesting access tokens for additional scopes.
This version creates the request to Google and attempts to handle
the callback by using what plain Django idioms.  The problem
is that the parameter string in the callback that Google issues
begins with a hash ('#'), and makes it difficult to see the
parameters in the <WSGIRequest: GET '/account/add_scope_callback/'>
object.