CSL-1412: WIP: Add logic to redirect to Bookshare login as necessary
When a user clicks an "Add to Bookshare" button, a check is made to see
if they are connected to Bookshare.  If they are not, they are redirected
to the Bookshare login workflow, using `/bookshare/search' as a `next`
parameter.  After the login is complete, the user will end up at
`/bookshare/search`, i.e. the Bookshare search page.