CSL-1647:  Implement two ways to retrieve organizational status
This adds the ability to retrieve the membership list associated with the
Bookshare account and the user's preferences, the latter using the v1 API.
The success or lack thereof when retrieving the membership list is used to
determine if the account is an individual (no membership list) or an
organizational account.  The user preferences could also be used for this
purpose.  It remains to decide which technique to actually use.

In any case, membership list retrieval will be needed for organziational
accounts for the sponsor import wokflow.