CSL-1662: Add back listener for window `pageHide` event
Previously replaced listener for `pageHide` with document's
`visibilitychange` event, but the latter is triggered only when the
session page changes with the result that the system was missing page
page turns within the app.  The new listener for `visibilitychange`
remains in place.

Also, the listener for `pageHide` is now hooked into
`navigator.sendBeacon()` (like document `visibilitychange`), bypassing
the client side message queue component.