NOJIRA: Update Cypress dependencies
Adds a minor work-around to get Cypress preference tests to pass, due to issue with Cypress and `<iframe>` use.

`window.parent` gets rewritten by Cypress - setting `"modifyObstructiveCode": false` in `cypress.json` is usually the fix, but doing so is causing other errors.  Went with the typeof checks for a quick fix.
Refs:
- https://github.com/cypress-io/cypress/issues/5538
- https://on.cypress.io/configuration#modifyObstructiveCode

This could use a revisit at a later time.