CSL-1533: Replace deprecated `Performance.timing`, part I.
This is a hybrid solution that uses the PerformanceObserver with
PerformanceNavigationTiming, Perfomance.timing, and a static call to
PerformanceNavigationTiming to get the load time.

Also, improve console messages so as to easily find PageTiming messages
in the log.

Next:  rely on the PerformanceObserver with PerformanceNavigationTiming
alone, for the load time.