CSL-1533: Improve handling of negative active duration
If the total inactive time is greater than the duration the page was
viewed, try halving the inactive time.  If half the inactive time is still
greater than the duration, set the active time to the duration.

Also, console messages use the same prefix as that used in the branch
that uses the beacon API to send the PageEnd messages
(feature/CSL-1662-time-tracking-with-beacon-api).