CSL-1813:  Update older organizational Bookshare account info with its type
All that was stored in the database with respect to the organizational aspect
of the account was true/false -- true if it was an organizational account,
and false if it was an individual account.  Now the property has a value that
indicates if it is a Sponsor vs. Member Organziational account, or an
Individual account.