CSL-1880: first draft of resources feature. I am unhappy with the reality of two concrete classes (Book and EducatorResource) inheriting from an abstract class that has most of the fields.  BookVersion and Paradata currently both have ForeignKey connections to Book, and that cannot be changed to point to an abstract class. Both would have to have two ForeignKey fields, only one of which is used, or something awkward like that. Plus there are just lots of other places where there have to be two very similar bits of code to operate on the two very similar classes.
I’m thinking I have to go back and change it all so that a Resource is just a Book with some sort of is_resource flag
(and a bunch of optional fields).

Or, we need to add in one of the django extensions that explicitly handle polymorphism, but I am reluctant to add that
sort of complication.