CSL-2046: Use a `clusive_user` QuerySet for period roster; fix bug
The bug was that if the "student" name in the request url was the
teacher of the class, it would show up at the top of the "Student
Details" popup.  The fix was changing the filter for the QuerySet
to include only students and guests.  And, if a non-student is
passed in the request, it is treated as a student that is not in
the period.