CSL-2046: Refactor to put check for `None` student in one place
When the student in the request is not in this class/period, set the
various panel data values to "None", zero, or whatever is appropriate
for that panel to indicate that there is no data for it.  This is done
once in an `except` clause rather than checking for a `None` student
for each panel.