CSL-2104: Add management command to check the word lookup lists
This compares the `words_looked_up` fields of a group of ParadataDaily
records with their parent Paradata.  It combines the looked-up words for
all ParadataDaily records into a single set and then compares that union
with the parent Paradata's `words_looked_up`.  They should match.  If they
do not match, an error message is printed.